refactor(i18n): centralize Locale type in types.ts
- Moved Locale type from locales.ts to types.ts for improved organization
- Introduced UnsanitizedLocaleParams type for handling unvalidated locale parameters
- Updated imports in getBlogPostData.ts, layout.tsx, NavigationLanguageToggle.tsx, and SiteLayout.tsx to use centralized types
