+++ apps/filecoin-site/src/app/_components/SiteLayout.tsx
@@ -9,7 +9,8 @@ import { ORGANIZATION_SCHEMA_BASE } from '@/constants/structuredDataConstants'
 
 import { Footer } from '@/components/Footer/Footer'
 
-import { DEFAULT_LOCALE, type Locale } from '@/i18n/locales'
+import { DEFAULT_LOCALE } from '@/i18n/locales'
+import type { Locale } from '@/i18n/types'
 
 // funnelSans also has an italic version in the same directory. We don't use italics as of now so it's not imported here.
 const funnelSans = localFont({
