+++ apps/filecoin-site/src/i18n/locales.ts
@@ -1,4 +1,2 @@
-export type Locale = (typeof LOCALES)[number]
-
 export const LOCALES = ['en', 'zh-cn'] as const
 export const DEFAULT_LOCALE = LOCALES[0]
