+++ apps/filecoin-site/src/i18n/types.ts
@@ -1,5 +1,5 @@
-import type { Locale } from './locales'
+import { LOCALES } from './locales'
 
-export type LocaleParams = {
-  locale: Locale
-}
+export type Locale = (typeof LOCALES)[number]
+export type LocaleParams = { locale: Locale }
+export type UnsanitizedLocaleParams = { locale: string }
